\name{facmodTS-package}
\alias{facmodTS-package}
\alias{facmodTS}
\docType{package}
\title{
\packageTitle{facmodTS}
}
\description{
\packageDescription{facmodTS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{facmodTS}
\packageIndices{facmodTS}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{facmodTS}

Maintainer: \packageMaintainer{facmodTS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
