% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitReturnsToFF3model.R
\name{fitReturnsToFF3model}
\alias{fitReturnsToFF3model}
\title{Fit Returns to FF3 Model}
\usage{
fitReturnsToFF3model(returns, datFF, digits = 2, title = NULL)
}
\arguments{
\item{returns}{Univariate xts returns data set}

\item{datFF}{Trivariate xts FF3 data set}

\item{digits}{Integer number of significant digits, with default digits = 2}

\item{title}{Character optional plot title, default = NULL}
}
\value{
List containing data frames LSRobFit and pvals, and a fit.models 
object fitsRobAndLS
}
\description{
Fits univariate xts returns to a Fama-French 3-factor data
set consisting of market excess returns (MKT), the small-minus-big (SMB)
 factor, and the high-minus low (HML) factor
}
\details{
The LSRobFit data frame contains the Alpha (intercept) MKT, SMB, HML
least squares and mOPt robust coefficient estimates, with their t-values in
parentheses, and the LS and robust adjusted R-squared values (AdjRSQ).  The
p-values data frame contains p-values of significance tests of robust versus
LS estimates for the overall model, and for the MKT, SMB, and HML factors
separately. The significance tests are computed with the lsRobTestMM.R
function in the PCRA package.
}
\examples{
args(fitReturnsToFF3model)
}
