% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitReturnsToFF4model.R
\name{fitReturnsToFF4model}
\alias{fitReturnsToFF4model}
\title{Fit Returns to FF4 Model}
\usage{
fitReturnsToFF4model(returns, datFF, digits = 2, title = NULL)
}
\arguments{
\item{returns}{Univariate xts returns data set}

\item{datFF}{Multivariate xts FF4 data set}

\item{digits}{Integer number of significant digits, with default digits = 2}

\item{title}{Character optional plot title, default = NULL}
}
\value{
List containing data frames LSRobFit and pvals, and a fit.models 
object fitsRobAndLS
}
\description{
Fits univariate xts returns to a Fama-French 4-factor data
set consisting of market excess returns (MKT), the small-minus-big (SMB)
 factor, the high-minus low (HML) factor, and the momentum (MOM) factor.
}
\details{
The LSRobFit data frame contains the Alpha (intercept) MKT, SMB, HML,
MOM least squares and mOPt robust coefficient estimates, with their t-values
in parentheses, and the LS and robust adjusted R-squared values (AdjRSQ).  The
p-values data frame contains p-values of significance tests of robust versus
LS estimates for the overall model, and for the MKT, SMB, HML and MOM factors
separately. The significance tests are computed with the lsRobTestMM.R
function in the PCRA package.
}
\examples{
args(fitReturnsToFF4model)
}
