% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTsfmSingleFactorStats.R
\name{fitTsfmSingleFactorStats}
\alias{fitTsfmSingleFactorStats}
\title{Title Fitted TSFM Single Factor Model Statistics}
\usage{
fitTsfmSingleFactorStats(fit, digits = 2)
}
\arguments{
\item{fit}{A fitted TSFM for a single factor}

\item{digits}{Integer number of digits for rounding, with default digits = 2}
}
\value{
A fitted TSFM object for a single factor
}
\description{
Computes fitted time series single factor model t-statistics, and 
adjusted R-squared values
}
\examples{
 args(fitTsfmSingleFactorStats)
}
