% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTsfmStats.R
\name{fitTsfmStats}
\alias{fitTsfmStats}
\title{Fitted TSFM Model Statistics}
\usage{
fitTsfmStats(fit, digits = 2)
}
\arguments{
\item{fit}{A TSFM fitted model object}

\item{digits}{Integer number of digits for rounding, with default digits = 2}
}
\value{
A fitted TSFM object
}
\description{
Computes fitted time series factor model (TSFM) models standard errors,
t-statistics, R-squared and adjusted R-squared values
}
\examples{
args(fitTsfmStats)
}
