% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facmodTS-package.R
\docType{package}
\name{facmodTS-package}
\alias{facmodTS}
\alias{facmodTS-package}
\title{facmodTS: Time Series Factor Models for Asset Returns}
\description{
Asset returns linear time series factor model fitting with both least squares and robust regression; model selection methods; returns factor and residuals decompositions; Standard Deviation, Value-at-Risk, and Expected Shortfall risk decompositions; tabular and graphical risk and performance evaluation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robustport/facmodTS}
}

}
\author{
\strong{Maintainer}: Doug Martin \email{martinrd3d@gmail.com}

Authors:
\itemize{
  \item Eric Zivot
  \item Sangeetha Srinivasan
}

Other contributors:
\itemize{
  \item Avinash Acharya [contributor]
  \item Yi-An Chen [contributor]
  \item Kirk Li [contributor]
  \item Lingjie Yi [contributor]
  \item Justin Shea [contributor]
  \item Mido Shammaa [contributor]
  \item Jon Spinney [contributor]
}

}
\keyword{internal}
