% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facmodTS-package.R
\docType{package}
\name{facmodTS-package}
\alias{facmodTS}
\alias{facmodTS-package}
\title{facmodTS: Time Series Factor Models for Asset Returns}
\description{
Linear factor model fitting for asset returns (three major types- time series, fundamental and statistical factor models); related risk (volatility, VaR and ES) and performance attribution (factor-contributed vs idiosyncratic returns); tabular displays of risk and performance reports; factor model Monte Carlo.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robustport/facmodTS}
}

}
\keyword{internal}
