% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facmodTS-package.R
\docType{package}
\name{facmodTS-package}
\alias{facmodTS}
\alias{facmodTS-package}
\title{facmodTS: Time Series Models for Asset Returns}
\description{
Supports teaching methods of estimating and testing time series models for use in robust portfolio construction and analysis. Unique in providing not only classical least squares, but also modern robust model fitting methods which are not much influenced by outliers. Includes returns and risk decompositions, with user choice of standard deviation, value-at-risk, and expected shortfall risk measures. "Robust Statistics Theory and Methods (with R)", R. A. Maronna, R. D. Martin, V. J. Yohai, M. Salibian-Barrera (2019) \doi{10.1002/9781119214656}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robustport/facmodTS}
}

}
\author{
\strong{Maintainer}: Doug Martin \email{martinrd3d@gmail.com}

Authors:
\itemize{
  \item Eric Zivot
  \item Sangeetha Srinivasan
}

Other contributors:
\itemize{
  \item Avinash Acharya [contributor]
  \item Yi-An Chen [contributor]
  \item Kirk Li [contributor]
  \item Lingjie Yi [contributor]
  \item Justin Shea [contributor]
  \item Mido Shammaa [contributor]
  \item Jon Spinney [contributor]
}

}
\keyword{internal}
