% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facmodTS-package.R
\docType{package}
\name{facmodTS-package}
\alias{facmodTS}
\alias{facmodTS-package}
\title{facmodTS: Time Series Factor Models for Asset Returns}
\description{
The package is collection of functions that support teaching methods of estimating and testing time series factor models for use in robust portfolio construction and analysis. The fitting methods include both classical least squares, and modern robust model fitting methods which are not much influenced by outliers. The package functionality includes returns and risk decompositions of factor models into total, factors, and residuals components, with user choice of standard deviation, value-at-risk (VaR), or expected shortfall (ES) risk measures. The package is well-suited for use in a portfolio construction course, or a financial time series modeling and anlysis course, in a quantitative finance MS degree program.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robustport/facmodTS}
}

}
\author{
\strong{Maintainer}: Doug Martin \email{martinrd3d@gmail.com}

Authors:
\itemize{
  \item Eric Zivot
  \item Sangeetha Srinivasan
}

Other contributors:
\itemize{
  \item Avinash Acharya [contributor]
  \item Yi-An Chen [contributor]
  \item Kirk Li [contributor]
  \item Lingjie Yi [contributor]
  \item Justin Shea [contributor]
  \item Mido Shammaa [contributor]
  \item Jon Spinney [contributor]
}

}
\keyword{internal}
