% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facmodTS-package.R
\docType{package}
\name{facmodTS-package}
\alias{facmodTS}
\alias{facmodTS-package}
\title{facmodTS: Time Series Factor Models for Asset Returns}
\description{
Asset returns linear time series factor model fitting with both least squares and robust regression; model selection methods; returns factor and residuals decompositions; Standard Deviation, Value-at-Risk, and Expected Shortfall risk decompositions; tabular and graphical risk and performance evaluation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robustport/facmodTS}
}

}
\keyword{internal}
